Handle the case where the user is only in one group
Some users (e.g. the default user on a number of "cloud" images) are
only in one group. In that case, make-perms-test will fail with the
message:

Traceback (most recent call last):
  File "./make-perms-test", line 79, in <module>
    ''' % locals())
TypeError: %d format: a number is required, not NoneType

Which isn't very clear. This patch checks that a second group was
found, and errors out if not.

Signed-off-by: Matthew Vernon <mv3@sanger.ac.uk>