Generate top-level spec file from template.
The "charliecloud.spec" file in packaging/redhat/ is really only a
template -- its version is specified as "@VERSION@" rather than the
actual version -- but for some reason it lacks the ".in" extension.
This commit adds some Makefile-fu to generate a real spec file in the
top-level directory from the template.