tests: don't require version to contain Git hash (closes: #206)
In addition to the cited bug, requiring $CH_VERSION (derived from
"ch-run --version") is too strict. It means that one must rebuild
before running the tests every time, which is usually not needed.

We could also move this to common.bash, but that would further extend
the time consumption of the test suite, which is the opposite
direction from what we want (#186).